// packages/sync/src/api/generated/models/OnboardingMessage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Safely Sync
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OnboardingMessage
 */
export interface OnboardingMessage {
    /**
     * 32-byte hex-encoded Ed25519 public key
     * @type {string}
     * @memberof OnboardingMessage
     */
    newIdentityPubKey: string;
    /**
     * hex-encoded 32 bytes
     * @type {string}
     * @memberof OnboardingMessage
     */
    inviterEphemeralPubKey: string;
    /**
     * hex-encoded bytes
     * @type {string}
     * @memberof OnboardingMessage
     */
    ciphertext: string;
    /**
     * hex-encoded 24 bytes
     * @type {string}
     * @memberof OnboardingMessage
     */
    nonce: string;
    /**
     * hex-encoded 64 bytes
     * @type {string}
     * @memberof OnboardingMessage
     */
    signature: string;
}

/**
 * Check if a given object implements the OnboardingMessage interface.
 */
export function instanceOfOnboardingMessage(value: object): value is OnboardingMessage {
    if (!('newIdentityPubKey' in value) || value['newIdentityPubKey'] === undefined) return false;
    if (!('inviterEphemeralPubKey' in value) || value['inviterEphemeralPubKey'] === undefined) return false;
    if (!('ciphertext' in value) || value['ciphertext'] === undefined) return false;
    if (!('nonce' in value) || value['nonce'] === undefined) return false;
    if (!('signature' in value) || value['signature'] === undefined) return false;
    return true;
}

export function OnboardingMessageFromJSON(json: any): OnboardingMessage {
    return OnboardingMessageFromJSONTyped(json, false);
}

export function OnboardingMessageFromJSONTyped(json: any, ignoreDiscriminator: boolean): OnboardingMessage {
    if (json == null) {
        return json;
    }
    return {
        
        'newIdentityPubKey': json['newIdentityPubKey'],
        'inviterEphemeralPubKey': json['inviterEphemeralPubKey'],
        'ciphertext': json['ciphertext'],
        'nonce': json['nonce'],
        'signature': json['signature'],
    };
}

export function OnboardingMessageToJSON(json: any): OnboardingMessage {
    return OnboardingMessageToJSONTyped(json, false);
}

export function OnboardingMessageToJSONTyped(value?: OnboardingMessage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'newIdentityPubKey': value['newIdentityPubKey'],
        'inviterEphemeralPubKey': value['inviterEphemeralPubKey'],
        'ciphertext': value['ciphertext'],
        'nonce': value['nonce'],
        'signature': value['signature'],
    };
}