// packages/sync/src/api/generated/models/Snapshot.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Safely Sync
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Snapshot
 */
export interface Snapshot {
    /**
     * hex encoded bytes
     * @type {string}
     * @memberof Snapshot
     */
    ciphertext: string;
    /**
     * hex-encoded 24 bytes
     * @type {string}
     * @memberof Snapshot
     */
    nonce: string;
    /**
     * hex-encoded 16 bytes
     * @type {string}
     * @memberof Snapshot
     */
    kid: string;
    /**
     * hex-encoded 32 bytes
     * @type {string}
     * @memberof Snapshot
     */
    snapshotProof: string;
    /**
     * hex-encoded 64 bytes
     * @type {string}
     * @memberof Snapshot
     */
    signature: string;
}

/**
 * Check if a given object implements the Snapshot interface.
 */
export function instanceOfSnapshot(value: object): value is Snapshot {
    if (!('ciphertext' in value) || value['ciphertext'] === undefined) return false;
    if (!('nonce' in value) || value['nonce'] === undefined) return false;
    if (!('kid' in value) || value['kid'] === undefined) return false;
    if (!('snapshotProof' in value) || value['snapshotProof'] === undefined) return false;
    if (!('signature' in value) || value['signature'] === undefined) return false;
    return true;
}

export function SnapshotFromJSON(json: any): Snapshot {
    return SnapshotFromJSONTyped(json, false);
}

export function SnapshotFromJSONTyped(json: any, ignoreDiscriminator: boolean): Snapshot {
    if (json == null) {
        return json;
    }
    return {
        
        'ciphertext': json['ciphertext'],
        'nonce': json['nonce'],
        'kid': json['kid'],
        'snapshotProof': json['snapshotProof'],
        'signature': json['signature'],
    };
}

export function SnapshotToJSON(json: any): Snapshot {
    return SnapshotToJSONTyped(json, false);
}

export function SnapshotToJSONTyped(value?: Snapshot | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ciphertext': value['ciphertext'],
        'nonce': value['nonce'],
        'kid': value['kid'],
        'snapshotProof': value['snapshotProof'],
        'signature': value['signature'],
    };
}